Database connection established; configuration added

+++ WebShopBackend/WebShopBackend/Model/ItemOrder.cs
@@ -7,9 +7,9 @@ namespace WebShopBackend.Model
 {
     public class ItemOrder
     {
-        public long OrderId { get; set; }
+        public int OrderId { get; set; }
         public Order Order { get; set; }
-        public long ItemId { get; set; }
+        public int ItemId { get; set; }
         public Item Item { get; set; }
         public int Amount { get; set; }
         public double Price{ get; set; }
